pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "A modern Flask-based personal blog"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "flask-sqlalchemy>=3.1.0",
    "flask-migrate>=4.0.0",
    "flask-wtf>=1.2.0",
    "flask-login>=0.6.3",
    "authlib>=1.3.0",
    "requests>=2.31.0",
    "markdown>=3.5.0",
    "pygments>=2.17.0",
    "pillow>=10.0.0",
    "black>=25.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-flask>=1.3.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["app.py", "models.py", "config.py", "forms.py", "templates", "static"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
